Removed NavBar, added myCVs button, did some UI fixes and relocated component files.
