+++ react/src/components/Browse/NameList.js
