+++ react/src/components/Browse/RenamePopover.js
