+++ react/src/components/Editor/Editor.js
@@ -1,10 +1,9 @@
 import React, { Component } from 'react'
 import EditorButtonGroup from './EditorButtonGroup'
-import Sections from './Sections'
-import Preview from './Preview'
-import './css/Editor.css'
-import './css/NavBar.css'
-import { saveCV, loadCV } from './Api'
+import InputSections from './InputSections'
+import Preview from '../Preview'
+import './Editor.css'
+import { saveCV, loadCV } from '../Api'
 
 
 class Editor extends Component {
@@ -53,7 +52,7 @@ class Editor extends Component {
           />
         </div>
         <div className="sections">
-          <Sections
+          <InputSections
             text={this.state.text}
             updateText={text => this.updateText(text)}
           />
