+++ react/src/components/Header.css
@@ -1,0 +1,14 @@
+@font-face {
+	font-family: 'JungkaBold';
+	src: url('/fonts/Jungka-Bold.otf');
+}
+
+.header {
+	padding-bottom: 20px;
+	padding-top: 20px;
+	padding-left: 10px;
+	border-bottom: 1px solid #242730;
+	font-family: 'JungkaBold';
+	color: #FC6054;
+	font-size: 20px;
+}
