+++ react/src/components/Header.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Header() {
+  return (
+    <header className="header">
+      CV Editor
+    </header>
+  )
+}
+
+export default Header
