--- react/src/components/css/Browse.css
@@ -1,49 +1,0 @@
-.search {
-	float: left;
-	width: 20.5%;
-}
-
-#searchbutton {
-	float: left;
-	width: 4%;
-	min-width: 4%;
-}
-
-.browseSection {
-	float: left;
-	display: inline;
-	width: 25%;
-	padding-right: 5px;
-	padding-left: 5px;
-	font-family: 'JungkaRegular';
-}
-
-.list-group-item {
-    padding-left: 1em;
-    color: #242730;
-}
-
-.list-group-item.active {
-	background-color: #d5d7de;
-	border-color: #d5d7de;
-	color: #000000;
-}
-
-#namelist {
-	overflow: auto;
-	height: 80vh;
-}
-
-#cvlist {
-	overflow: auto;
-	height: 80vh;
-}
-
-.CVpreview {
-	overflow: auto;
-	height: 80vh;
-}
-
-html, body {
-  overflow: hidden;
-}
