--- react/src/components/css/NavBar.css
@@ -1,37 +1,0 @@
-@font-face {
-	font-family: 'JungkaBold';
-	src: url('/fonts/Jungka-Bold.otf');
-}
-
-.navigation {
-	padding-bottom: 20px;
-	padding-top: 20px;
-	padding-left: 10px;
-	border-bottom: 1px solid #242730;
-	font-family: 'JungkaBold';
-}
-
-.btn-secondary {
-	background-color: #fff;
-	border-color: #fff;
-	color: #FC6054;
-	font-size: 20px;
-}
-
-.btn-secondary:hover {
-	color: #ab463a;
-	background-color: #fff;
-	border-color: #fff;
-}
-
-.btn-secondary:not([disabled]):not(.disabled).active {
-	background-color: #FC6054;
-	color: #FFFFFF;
-	border-color: #FC6054;
-}
-
-.btn-secondary:not([disabled]):not(.disabled).active:hover {
-	background-color: #ab463a;
-	color: #fff;
-	border-color: #ab463a;
-}
