+++ react/src/index.css
@@ -25,11 +25,13 @@
 .button {
   color: #242730;
   font-family: 'JungkaRegular';
-
+  box-shadow: none;
 }
 
 .button:hover {
   background-color: #242730;
+  border-color: #242730;
+  box-shadow: none;
 }
 
 /* The right color stays when the Export dropdown is opened */
@@ -37,6 +39,7 @@
   background-color: #242730;
   border-color: #242730;
   color: #FFFFFF;
+  box-shadow: none;
 }
 
 .dropdown-menu {
@@ -51,6 +54,7 @@
   float: right;
   width: 100%;
   font-family: 'JungkaRegular';
+  box-shadow: none;
 }
 
 #editor-buttonheader {
@@ -60,6 +64,7 @@
 .exportgroup {
   float: right;
   font-family: 'JungkaRegular';
+  box-shadow: none;
 }
 
 .CVpreview {
@@ -70,29 +75,3 @@
   padding-left: 15px;
   overflow: auto;
 }
-
-/*CSS for the lines in between the containers*/
-
-.lineContainer {
-  width: 1px;
-  height: 100%;
-  left: 50%;
-  position: absolute;
-  top: 20%;
-}
-
- #lineContainer {
-  width: 1px;
-  height: 100vh;
-  left: 25%;
-  position: absolute;
-
-}
-
-.line {
-  position: absolute;
-  width:100%;
-  height: 100vh;
-  background: #d5d7de;
-}
-
